Update ek package to v12
