+++ README.md
@@ -28,7 +28,7 @@ Before the initial install allows git to use redirects for [pkg.re](https://gith
 git config --global http.https://pkg.re.followRedirects true
 ```
 
-Make sure you have a working Go 1.10+ workspace ([instructions](https://golang.org/doc/install)), then:
+Make sure you have a working Go 1.12+ workspace ([instructions](https://golang.org/doc/install)), then:
 
 ```
 go get github.com/essentialkaos/shdoc
@@ -43,6 +43,12 @@ go get -u github.com/essentialkaos/shdoc
 #### Prebuilt binaries
 
 You can download prebuilt binaries for Linux and OS X from [EK Apps Repository](https://apps.kaos.st/shdoc/latest).
+
+To install the latest prebuilt version of bibop, do:
+
+```bash
+bash <(curl -fsSL https://apps.kaos.st/get) shdoc
+```
 
 ### Command-line completion
